refactor: Landmark, LandmarkRaster 관련 객체의 id 타입 long으로 통일

+++ back/src/main/java/dev/group2/landmark_be/map/dto/response/LandmarkRasterResponse.java
@@ -5,8 +5,8 @@ import java.math.BigDecimal;
 import org.locationtech.jts.geom.Polygon;
 
 public record LandmarkRasterResponse(
-	Integer id,
-	Integer landmarkId,
+	Long id,
+	Long landmarkId,
 	String indexType,	// NDVI, NDMI
 	Integer year,
 	Integer month,
